Added way to enter weight in app

+++ app/src/main/java/edu/neu/madcourse/fastit/FastingSession.java
@@ -23,5 +23,5 @@ public class FastingSession {
     public String progressImagePath;
 
     @ColumnInfo
-    public double weight;
+    public float weight;
 }
